NPCs scripted for movement
although not quite finished, the NPCNavMesh script has been updated to find a random transform out of 9 possible transforms for each NPC. an NPC is a servant or not. If a servant, it must follow a noblewoman.

the part that isn't fixed is removing "taken" transforms from the list of possible transforms.