As alluded to in the previous commit, changed it from Ensembleobject to EnsembleObjectIntro.
Additional changes from the previous commit but not included in the previous commit message, I 'fake' the Omeka screen so that it is more in-line with with the other tutorials about Actions and History (this makes it not depend on an actual Omeka call). I wrote some placeholder text here to describe Omeka that probably should be revised. Also, I made it so that the ticket taker's halo shows up.